+++ CashierClasses/CashierClasses/Product.cs
@@ -12,9 +12,7 @@ namespace CashierClasses
         private Int32 qty;
 
         private double price;
-
-
-        private decimal vat;
+        private double vat;
 
         private string manufacturer;
         private string supplier;
@@ -26,7 +24,7 @@ namespace CashierClasses
 
 
 
-        public Product(int idy, double pricy, string manufac, string supply, string productG, decimal vaty, Boolean ppkg, Boolean ppHg, string namey)
+        public Product(int idy, double pricy, string manufac, string supply, string productG, double vaty, Boolean ppkg, Boolean ppHg, string namey)
         {
             Id = idy;
             Price = pricy;
@@ -47,7 +45,7 @@ namespace CashierClasses
         public string Manufacturer { get; set; }
         public string Supplier { get; set; }
         public string ProductGroup { get; set; }
-        public decimal Vat { get; set; }
+        public double Vat { get; set; }
         public Boolean PricePerKG { get; set; }
         public Boolean PricePerHG { get; set; }
         public string Name { get; set; }
